pyproject.toml
[project]
name = "entsoe-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "entsoe-py>=0.7.8",
    "pandas>=2.3.3",
    "pandera>=0.26.1",
    "polars>=1.35.2",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }
packages = ["src/entsoe_project"]
